ResponseEntity

+++ spring4mvc/src/main/java/io/spring4mvc/spring4mvc/user/UserApiController.java
@@ -3,6 +3,8 @@ package io.spring4mvc.spring4mvc.user;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -36,18 +38,37 @@ public class UserApiController {
 //	}
 //	http post http://localhost:8888/api/users email=ltks3@naver.com birthDate=2017-11-03 tastes:='["httpie","pies"]' not working
 	@PostMapping("/users")
-	public User createUser(@RequestBody User user) {
-		return userRepository.save(user);
+	public ResponseEntity<User> createUser(@RequestBody User user) {
+		HttpStatus status = HttpStatus.OK;
+		
+		if (!userRepository.exists(user.getEmail())) {
+			status = HttpStatus.CREATED;
+		}
+		
+		User savedUser = userRepository.save(user);
+		
+		return new ResponseEntity<>(savedUser, status);
 	}
 	
 	@PutMapping("/user/{email}")
-	public User updateUser(@PathVariable String email, @RequestBody User user) {
-		return userRepository.save(email, user);
+	public ResponseEntity<User> updateUser(@PathVariable String email, @RequestBody User user) {
+		if (!userRepository.exists(email)) {
+			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+		}
+		
+		User updtUser = userRepository.save(email, user);
+		
+		return new ResponseEntity<>(updtUser, HttpStatus.CREATED);
 	}
 	
 	@DeleteMapping("/user/{email}")
-	public void deleteUser(@PathVariable String email) {
+	public ResponseEntity<User> deleteUser(@PathVariable String email) {
+		if (!userRepository.exists(email)) {
+			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+		}
+		
 		userRepository.delete(email);
+		return new ResponseEntity<>(HttpStatus.OK);
 	}
